Update for Chicken bullet
The last commit details:

I created basketball zombie and wrote a C# script for him to shot bullet to player (ofc with bullet settings) and the sound effect of hitting it by player.

Also I created level portal for travel to next level.

I also created a level2 boss who is chicken zombie to shot giant chicken with sound effect.

For this update, I modified the chicken prefab and make it more idolized.